Add cli-template.clj for clojure scripting with lein-exec
This template is meant to be used as a starting point to make
clojure command line interface scripts to use with the lein
exec plugin. It demonstrates the basics for pulling in a
string and integer option.

TODO: Could update this in order to handle errors and more
complexity, but for now just wanted to capture it so that I
could use it as a starting point