Update lein exec template to allow dynamic m2 loading
I encountered a blog post that gave some good pointers for
using lein exec

http://charsequence.blogspot.in/2012/04/scripting-clojure-with-leiningen-2.html

The first piece of advice was to use #! /bin/sh lein exec so that
with chmod a+x, you can run the script directly with ./clj-template.clj.

Next, I worked on a Wikipedia web scraper, and want to use enlive in my
script. In order to do that, I had to use the leiningen.exec/deps call, which
is also added in this commit.

With these changes, it should be really easy to create clojure scripts to
use for any problems I encounter in the wild.