Add mode variable to open clj-template in emacs clojure-mode
Found a good resource describing how emacs determines the mode to use
for opening files.

https://www.gnu.org/software/emacs/manual/html_node/emacs/Choosing-Modes.html

I ran into a problem with clj-template being opened in shell-mode because
of the first line being #! /bin/sh

The second line of the template now contains a file-local mode variable
to explicitly tell emacs to open with clojure-mode.